Update tests and bug fixes (#237)
* Updated failing syntax tests in various directories {array, globalFunctions,
immutable, inline_arrays}. Updated failing-exec-tests.

* Bug fix: In handling of rational numbers in shift/exp in ConstantEvaluator.cpp and Types.cpp.

* Updated failing syntax tests in various directories {abiEncoder, constructor,
dataLocations, denominations, functionCalls, types}. Updated failing execution tests.

* Updated failing execution tests: semanticTests/{constants, operators}.

* Updated failing execution tests: semanticTests/{enums, uninitializedFunctionPointer}.

* Updated failing execution tests: semanticTests/{getters, immutable}.

* Updated failing execution tests: semanticTests/{constructor, functionTypes, payable}.

* Updated failing compilation tests.

* Updated failing execution tests: semanticTests/{storage, reverts, functionCall}.

+++ test/libsolidity/semanticTests/functionTypes/pass_function_types_externally.sol
@@ -1,9 +1,9 @@
 contract C {
-    function f(uint256 x) public returns (uint256) {
+    function f(uint x) public returns (uint) {
         return this.eval(this.g, x);
     }
 
-    function f2(uint256 x) public returns (uint256) {
+    function f2(uint x) public returns (uint) {
         return eval(this.g, x);
     }
 
@@ -11,7 +11,7 @@ contract C {
         return x(a);
     }
 
-    function g(uint256 x) public returns (uint256) {
+    function g(uint x) public returns (uint) {
         return x + 1;
     }
 }
@@ -19,5 +19,5 @@ contract C {
 // ====
 // compileViaYul: also
 // ----
-// f(uint256): 7 -> 8
-// f2(uint256): 7 -> 8
+// f(uint): 7 -> 8
+// f2(uint): 7 -> 8
